Added sorting by `tags` and improved on the alphanumeric sorting, also clean up all those sorting utilities in one file
