+++ packages/webapp/src/components/lists/ContactList/columns.tsx
@@ -13,8 +13,7 @@ import { GenderText, getGenderText } from './GenderText'
 import { RaceText, getRaceText } from './RaceText'
 import { useLocale } from '~hooks/useLocale'
 import { Namespace, useTranslation } from '~hooks/useTranslation'
-import { sortByAlphanumeric } from '~utils/sortByAlphanumeric'
-import { sortByDate } from '~utils/sortByDate'
+import { sortByAlphanumeric, sortByDate } from '~utils/sorting'
 
 export function usePageColumns(actions: IMultiActionButtons<Contact>[]): IPaginatedListColumn[] {
 	const { t } = useTranslation(Namespace.Clients)
@@ -45,7 +44,7 @@ export function usePageColumns(actions: IMultiActionButtons<Contact>[]): IPagina
 				isSortable: true,
 				sortingFunction: sortByDate,
 				sortingValue(contact: Contact) {
-					return { date: contact.dateOfBirth } // See '~utils/sortByDate'
+					return { date: contact.dateOfBirth } // See '~utils/sorting'
 				}
 			},
 			{
