+++ packages/webapp/src/hooks/api/useInactiveEngagementList.ts
@@ -10,7 +10,7 @@ import { get } from 'lodash'
 import { useRecoilState } from 'recoil'
 import { inactiveEngagementListState } from '~store'
 import { useEffect } from 'react'
-import { sortByDate } from '~utils/sortByDate'
+import { sortByDate } from '~utils/sorting'
 import { Namespace, useTranslation } from '~hooks/useTranslation'
 import { SUBSCRIBE_TO_ORG_ENGAGEMENTS } from './useEngagementList/useEngagementSubscription'
 import { createLogger } from '~utils/createLogger'
