+++ packages/webapp/src/types/Sorting.ts
@@ -3,17 +3,24 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 
-import type { HasDate } from '~utils/sortByDate'
-import type { Alphanumeric } from '~utils/sortByAlphanumeric'
+import { Tag } from '@cbosuite/schema/dist/client-types'
+
+export type Alphanumeric = string | number
+
+export type HasDate = {
+	date: string
+}
+
+export type Tags = Tag[]
 
 export enum SortingOrder {
 	DESC = -1,
 	ASC = 1
 }
 
-export type SortingFunction = (a: SortingInput, b: SortingInput, order: SortingOrder) => number
+export type SortingInput = string | number | Alphanumeric | Date | HasDate | Tags
 
-export type SortingInput = string | number | Alphanumeric | Date | HasDate
+export type SortingFunction = (a: SortingInput, b: SortingInput, order: SortingOrder) => number
 
 export type SortingValue = (x: Record<string, unknown>) => SortingInput
